requirements.txt
vanna[google,sqlite]>=2.0.0
fastapi
uvicorn[standard]
python-dotenv
google-genai
pandas
plotly
